Removed MainFavoritasFragment
Now its a view inside MainPageFragment.
To avoid having too much noise there I've extracted all the View related stuff to a different class.
