+++ SeviBus/src/main/java/com/sloy/sevibus/ui/fragments/MainPageFragment.java
@@ -1,9 +1,11 @@
 package com.sloy.sevibus.ui.fragments;
 
 import android.content.Context;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
+import android.support.annotation.Nullable;
 import android.support.design.widget.Snackbar;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
@@ -17,22 +19,25 @@ import android.view.ViewGroup;
 
 import com.sloy.sevibus.BuildConfig;
 import com.sloy.sevibus.R;
+import com.sloy.sevibus.resources.StuffProvider;
 import com.sloy.sevibus.ui.activities.BusquedaActivity;
 import com.sloy.sevibus.ui.activities.PreferenciasActivity;
-import com.sloy.sevibus.ui.fragments.main.FavoritasMainFragment;
 import com.sloy.sevibus.ui.fragments.main.LineasCercanasMainFragment;
 import com.sloy.sevibus.ui.fragments.main.ParadasCercanasMainFragment;
+import com.sloy.sevibus.ui.mvp.presenter.FavoritasMainPresenter;
+import com.sloy.sevibus.ui.mvp.view.FavoritasMainViewContainer;
 
 import de.cketti.library.changelog.ChangeLog;
 
 public class MainPageFragment extends BaseDBFragment {
 
-    private static final String FRAG_FAVORITAS = "f_favoritas";
     private static final String FRAG_PARADAS_CERCANAS = "f_p_cercanas";
     private static final String FRAG_LINEAS_CERCANAS = "f_l_cercanas";
 
     private static final String PREF_SHOW_NEW_VERSION_LATEST_SEEN = "newversion_last_seen";
     private static final int NEW_VERSION_SNACKBAR_DURATION = 10000;
+
+    private FavoritasMainPresenter favoritasPresenter;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -40,10 +45,35 @@ public class MainPageFragment extends BaseDBFragment {
         setHasOptionsMenu(true);
     }
 
+
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        favoritasPresenter = new FavoritasMainPresenter(StuffProvider.getObtainFavoritasAction(getActivity()));
+    }
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View v = inflater.inflate(R.layout.fragment_main_home, container, false);
         return v;
+    }
+
+    @Override
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        favoritasPresenter.initialize(new FavoritasMainViewContainer(favoritasPresenter, view.findViewById(R.id.fragment_main_favoritas)));
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        favoritasPresenter.update();
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        favoritasPresenter.pause();
     }
 
     @Override
@@ -55,7 +85,6 @@ public class MainPageFragment extends BaseDBFragment {
     private void setupUI() {
         FragmentManager fm = getChildFragmentManager();
         FragmentTransaction trans = fm.beginTransaction();
-        setupFavoritas(fm, trans);
         setupParadasCercanas(fm, trans);
         setupLineasCercanas(fm, trans);
         trans.commit();
@@ -86,18 +115,6 @@ public class MainPageFragment extends BaseDBFragment {
         }
 
         prefs.edit().putInt(PREF_SHOW_NEW_VERSION_LATEST_SEEN, currentVersion).apply();
-    }
-
-    private void setupFavoritas(FragmentManager fm, FragmentTransaction trans) {
-        Fragment f = fm.findFragmentByTag(FRAG_FAVORITAS);
-        if (f == null) {
-            f = FavoritasMainFragment.getInstance();
-        }
-        if (f.isAdded()) {
-            trans.attach(f);
-        } else {
-            trans.add(R.id.fragment_main_favoritas, f, FRAG_FAVORITAS);
-        }
     }
 
     private void setupParadasCercanas(FragmentManager fm, FragmentTransaction trans) {
